提交内容(除评估标准内容外，有额外的内容)/src/back_end_and_database/your_project_name/requirements.txt
asgiref==3.8.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
psycopg2==2.9.9
setuptools==69.5.1
Django
wheel==0.43.0
sqlparse
tzdata